+++ server/src/server.ts
@@ -42,7 +42,7 @@ import {
 	FileChangeType,
 	Proposed,
 	WorkspaceFolder
-} from 'vscode-languageserver';
+} from 'vscode-languageserver/node';
 
 import {
 	TextDocument,